pyproject.toml
[tool.poetry]
name = "drone-rocket-finder-ui"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.14.1"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
opencv-python = "^4.10.0.84"
streamlit = "^1.36.0"
streamlit-jupyter = "^0.2.1"
streamlit-image-zoom = "^0.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
